requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
db-dtypes==1.4.3
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
executing==2.2.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
future==1.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.21.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.36.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.8
google-cloud-pubsub==2.29.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-genai==1.10.0
google-generativeai==0.3.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.63.0rc1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
Markdown==3.4.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
oauthlib==3.2.2
#opencv-python==4.11.0.86
opencv-python-headless
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas==2.2.3
pandas-gbq==0.28.1
parso==0.8.4
pexpect==4.9.0
pillow
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydata-google-auth==1.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
virtualenv==20.31.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug
wrapt==1.17.2
zipp==3.21.0
